requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
google-generativeai
python-dotenv
tqdm
